+++ spb/backends/base_backend.py
@@ -279,8 +279,6 @@ class Plot:
                 if ((not ("process_piecewise" in kwargs.keys())) and
                     ((_legend is True) or (_legend is None))):
                     self.legend = True
-            if self.legend and (len([s for s in self._series if s.is_3Dsurface and not s.use_cm]) > 1):
-                self.legend = False
 
         # Objects used to render/display the plots, which depends on the
         # plotting library.